requirements.txt
# requirements.txt

fastapi
uvicorn
python-dotenv
python-multipart
sqlalchemy
psycopg2-binary
pydantic
google-generativeai
PyPDF2
pillow
